bot/requirements.txt
aiogram==2.25.1
APScheduler==3.10.1
python-dotenv==1.0.0